Add missing piece for calling summarise activity
