+++ .ebstalk.apps.env/webapp.env
@@ -64,6 +64,7 @@ GITHUB_APP_CLIENT_SECRET="{{pull:secretsmanager:/io.cv.app/prd/github:SecretStri
 GITHUB_APP_WEBHOOK_SECRET="{{pull:secretsmanager:/io.cv.app/prd/github:SecretString:github_app_webhook_secret}}"
 GITHUB_ACCESS_TOKEN="{{pull:secretsmanager:/io.cv.app/prd/github:SecretString:github_access_token}}"
 OPENAI_API_KEY="{{pull:secretsmanager:/io.cv.app/prd/openai:SecretString:openai_api_key}}"
+CONNECTOR_KEY="{{pull:secretsmanager:/io.cv.app/prd/charmverse:SecretString:connector_key}}"
 
 # Frontend Secrets Start Here
 REACT_APP_APP_ENV="production"
