+++ pages/api/github/summarise-activity.ts
@@ -1,0 +1,32 @@
+import type { NextApiRequest, NextApiResponse } from 'next';
+import nc from 'next-connect';
+
+import { summariseDeveloperActivity } from 'lib/github/summariseDeveloperActivity';
+import type { PullRequestSummaryWithFilePatches } from 'lib/github/summarisePullRequest';
+import { onError, onNoMatch, requireKeys } from 'lib/middleware';
+import { requireConnectorKey } from 'lib/middleware/requireConnectorKey';
+
+const handler = nc({
+  onError,
+  onNoMatch
+});
+
+handler
+  .use(requireConnectorKey)
+  .use(requireKeys(['githubUsername'], 'body'))
+  .post(summariseDeveloperActivityController);
+
+async function summariseDeveloperActivityController(
+  req: NextApiRequest,
+  res: NextApiResponse<PullRequestSummaryWithFilePatches[]>
+) {
+  const summaries = await summariseDeveloperActivity({
+    githubUsername: req.body.githubUsername,
+    limit: req.body.limit,
+    fromDate: req.body.fromDate
+  });
+
+  res.status(200).json(summaries);
+}
+
+export default handler;
